AST is hierarchy of nodes
As is I don't think the explanation of the AST is clear as to what it is or why it is needed.

Also I think it is important to explain what an AST not what it isn't, which is what the as written N.B. is doing.

I think the overall explanations  and examples in this chapter are very good but the intro into what the purpose of the explanations.  is needed. You need to give the reader something to hand their had on before going into details.

P.S. if you think my comments a "off the wall" just let me know and I'll stick to typos :-)